refactor(google-maps): upgrade google-maps and consolidate logic
