+++ package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.47",
-    "@google/maps": "^0.3.1",
+    "@google/maps": "^0.4.6",
     "cli-table2": "^0.2.0",
     "colors": "^1.1.2 ",
     "commander": "^2.9.0",
